\name{lazy.frame-package}
\alias{lazy.frame-package}
\title{File Frames: A lazy person's file-backed data frame.}
\description{
Lazy frames are simple relatives of data frames, backed by text files. Backing
files are parsed and converted to data frames on demand.  Lazy frames are
useful for quickly and efficiently extracting subsets from large csv and other
text files without loading the entire file. They support normal and
gzip-compressed files.
}
